# Services/Bones.Api/OpenApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Bones API",
    "description": "Its an API and it does stuff",
    "contact": {
      "name": "GitHub Issues",
      "url": "https://github.com/CorruptComputer/Bones/issues"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/CorruptComputer/Bones/blob/develop/LICENSE"
    },
    "version": "develop"
  },
  "paths": {
...
    "/Item/fields/{fieldId}/latest": {
      "get": {
        "tags": [
          "Item"
        ],
        "summary": "Gets the latest version of a field",
        "operationId": "GetLatestItemFieldVersionAsync",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "description": "The ID of the field",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLatestItemFieldVersionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Item/layouts/{layoutId}/latest": {
      "get": {
        "tags": [
          "Item"
        ],
        "summary": "Get an ItemLayoutVersion by its LayoutID and version, or the latest version if no version is specified",
        "operationId": "GetLatestItemLayoutVersionAsync",
        "parameters": [
          {
            "name": "layoutId",
            "in": "path",
            "description": "The ID of the layout",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetItemLayoutVersionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Item/layouts/{layoutId}/{requestedVersion}": {
      "get": {
        "tags": [
          "Item"
        ],
        "summary": "Get an ItemLayoutVersion by its LayoutID and version, or the latest version if no version is specified",
        "operationId": "GetItemLayoutVersionAsync",
        "parameters": [
          {
            "name": "layoutId",
            "in": "path",
            "description": "The ID of the layout",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestedVersion",
            "in": "path",
            "description": "The version if you want a specific version instead of the latest",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetItemLayoutVersionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Item/layouts/{layoutId}/latest/fields": {
      "get": {
        "tags": [
          "Item"
        ],
        "summary": "Gets the latest version of a layout",
        "operationId": "GetLatestItemLayoutVersionFieldsAsync",
        "parameters": [
          {
            "name": "layoutId",
            "in": "path",
            "description": "The ID of the layout",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetLatestItemLayoutVersionFieldsResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressFields": {
        "enum": [
          "none",
          "streetNumber",
          "streetName",
          "cityOrPlace",
          "stateOrProvince",
          "postalCode",
          "county",
          "country"
        ],
        "type": "string"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringStringListKeyValuePair"
            },
            "description": "The errors that occurred, with the key being either the input that was invalid and the list of reasons it was invalid, or 'server' with the list of server errors.",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "description": "The error message, if any.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The response body is empty, this is a workaround for the limitations of the API client."
      },
      "FieldType": {
        "enum": [
          "textField",
          "textBox",
          "integer",
          "decimal",
          "boolean",
          "dateTime",
          "valueList",
          "geoLocation"
        ],
        "type": "string"
      },
      "GeoLocationType": {
        "enum": [
          "osmObject",
          "latLon",
          "address",
          "geometry"
        ],
        "type": "string"
      },
...
      "GetItemLayoutVersionResponse": {
        "required": [
          "fieldVersions",
          "friendlyIdPrefix",
          "latestVersion",
          "layoutUse",
          "name",
          "projectId",
          "version",
          "versionUpdateAvailable"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the item layout"
          },
          "projectId": {
            "type": "string",
            "description": "The ID of the project this layout belongs to",
            "format": "uuid"
          },
          "layoutUse": {
            "$ref": "#/components/schemas/ItemLayoutUse"
          },
          "friendlyIdPrefix": {
            "maxLength": 6,
            "type": "string",
            "description": "The prefix at the start of a Friendly ID for items using this layout, up to 6 letters."
          },
          "version": {
            "type": "integer",
            "description": "The version number for this ItemLayoutVersion",
            "format": "int64"
          },
          "latestVersion": {
            "type": "integer",
            "description": "The current latest version of this layout",
            "format": "int64"
          },
          "versionUpdateAvailable": {
            "type": "boolean",
            "description": "Is there a version update available for this layout version?",
            "readOnly": true
          },
          "fieldVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Int32GuidKeyValuePair"
            },
            "description": "The field versions"
          }
        },
        "additionalProperties": false,
        "description": "API response for the GetLatestItemLayoutVersionAsync endpoint"
      },
      "GetLatestItemFieldVersionResponse": {
        "required": [
          "fieldId",
          "fieldVersionId",
          "isRequired",
          "name",
          "type",
          "version"
        ],
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string",
            "description": "ID of the item field",
            "format": "uuid"
          },
          "fieldVersionId": {
            "type": "string",
            "description": "ID of the item field version",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "description": "Version number of the item field",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Name of the item field"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Is the field required?"
          },
          "type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "canBeNegative": {
            "type": "boolean",
            "description": "If the field is numeric, can it be negative?",
            "nullable": true
          },
          "possibleValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringStringValueMatchingTypeKeyValuePair"
            },
            "description": "If the field is a ValueList, the possible values",
            "nullable": true
          },
          "geoLocationType": {
            "$ref": "#/components/schemas/GeoLocationType"
          },
          "requiredAddressFields": {
            "$ref": "#/components/schemas/AddressFields"
          }
        },
        "additionalProperties": false,
        "description": "API response for the GetLatestItemFieldVersionAsync endpoint"
      },
      "GetLatestItemLayoutVersionFieldsResponse": {
        "required": [
          "id",
          "isRequired",
          "name",
          "orderNumber",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID for the ItemFieldVersion",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "integer",
            "description": "The order number for which this field should be displayed",
            "format": "int32"
          },
          "name": {
            "maxLength": 512,
            "type": "string",
            "description": "The name of this ItemFieldVersion"
          },
          "type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Is this field required to have a value?"
          },
          "possibleValues": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "If the field type is a ValueList, this will contain the possible values for this field.",
            "nullable": true
          },
          "canBeNegative": {
            "type": "boolean",
            "description": "If the Type of this field is either an Integer or Decimal, can it be negative?",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "API response for the GetLatestItemLayoutVersionFieldsAsync endpoint"
      },
...
      "Int32GuidKeyValuePair": {
        "required": [
          "key",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ItemLayoutUse": {
        "enum": [
          "none",
          "workItems",
          "assets"
        ],
        "type": "string"
      },
...
      "StringStringListKeyValuePair": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringStringValueMatchingTypeKeyValuePair": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "$ref": "#/components/schemas/StringValueMatchingType"
          }
        },
        "additionalProperties": false
      },
      "StringValueMatchingType": {
        "enum": [
          "exact",
          "caseInvariant",
          "soundex"
        ],
        "type": "string"
      },
...
    }
  }
}